test: Move fixtures into src folder
