+++ src/__fixtures__/index.js
