+++ src/components/Video/Video.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { create } from 'react-test-renderer';
 
-import { video } from '../../../__fixtures__';
+import { video } from '../../__fixtures__';
 import Video from './Video';
 
 describe('Video', () => {
